move hcn code to file with windows build tag
