+++ npm/pkg/dataplane/types_linux.go
@@ -1,0 +1,4 @@
+package dataplane
+
+// npmEndpoint holds info relevant for endpoints in windows
+type npmEndpoint struct{}
